+++ src/main/java/webserver/controller/Controller.java
@@ -1,8 +1,7 @@
 package webserver.controller;
 
 import webserver.httpUtils.Request;
-import webserver.httpUtils.Response;
-import webserver.httpUtils.ResponseHandler;
+import webserver.httpUtils.ResponseSender;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -10,5 +9,5 @@ import java.io.OutputStream;
 public interface Controller {
     public void exec(Request req, OutputStream out) throws IOException;
 
-    ResponseHandler resHandler = null;
+    ResponseSender resHandler = null;
 }
