+++ src/main/java/webserver/controller/StaticFileController.java
@@ -2,7 +2,7 @@ package webserver.controller;
 
 import webserver.httpUtils.Request;
 import webserver.httpUtils.Response;
-import webserver.httpUtils.ResponseHandler;
+import webserver.httpUtils.ResponseSender;
 import webserver.service.LogInService;
 import webserver.service.Service;
 import webserver.service.SignUpService;
@@ -13,23 +13,23 @@ import java.io.OutputStream;
 public class StaticFileController implements Controller{
     public StaticFileController(){}
 
-    private ResponseHandler resHandler;
+    private ResponseSender resHandler;
 
     @Override
     public void exec(Request req, OutputStream out) throws IOException {
         String query = req.getReqLine().getQuery();
         Service service = new Service() {};
-        if(query.contains("user/create"))
+        if(query.endsWith("user/create"))
         {
             service = new SignUpService();
         }
-        if(query.contains("user/login"))
+        if(query.endsWith("user/login"))
         {
             service = new LogInService();
         }
 
         Response res = service.exec(req);
-        resHandler = new ResponseHandler(res);
+        resHandler = new ResponseSender(res);
         resHandler.sendRes(out);
     }
 }
